seperated kobold logic from renderer
